# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: Код ошибки, например, 400 или 500
        type: integer
      message:
        description: Сообщение об ошибке
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
  models.Order:
    properties:
      order_id:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.OrderProduct'
        type: array
      user_id:
        type: integer
    type: object
  models.OrderProduct:
    properties:
      order_id:
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  models.OrderResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      name:
        type: string
      price:
        type: number
      rating:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
    url: http://example.com/support
  description: API для интернет-магазина спортивного питания. Содержит функционал
    для управления пользователями, продуктами, заказами и отзывами.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Sports Nutrition Store API
  version: "1.0"
paths:
  /admin/orders:
    get:
      consumes:
      - application/json
      description: Возвращает список заказов, включая информацию о продуктах в заказах
      parameters:
      - description: Токен доступа пользователя (JWT)
        in: header
        name: Authorization
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        name: limit
        type: integer
      - default: id
        description: Поле для сортировки
        in: query
        name: sort
        type: string
      - default: asc
        description: Направление сортировки
        in: query
        name: order
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: ID заказа
        in: query
        name: order_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заказов с продуктами
          schema:
            items:
              $ref: '#/definitions/models.OrderResponse'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка всех заказов
      tags:
      - orders
  /admin/orders/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет указанный заказ вместе с привязанными продуктами.
      parameters:
      - description: Токен пользователя
        in: header
        name: Authorization
        type: string
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление заказа
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка на сервере
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление заказа
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...